refactor(29953): refactor the function fetcher
