+++ hivemq-edge-frontend/src/extensions/datahub/hooks/useGetFilteredFunctions.tsx
@@ -1,43 +1,16 @@
 import { useMemo } from 'react'
 
-import { BehaviorPolicyTransitionEvent, type FunctionSpecs } from '@/api/__generated__'
+import type { BehaviorPolicyTransitionEvent } from '@/api/__generated__'
 import { DataHubNodeType } from '@datahub/types.ts'
 import { useGetAllFunctionSpecs } from '@datahub/api/hooks/DataHubFunctionsService/useGetAllFunctionSpecs.ts'
+import { MqttTransformFunction } from '@datahub/hooks/useFilteredFunctionsFetcher.tsx'
 
-export const MqttTransformFunction: FunctionSpecs = {
-  functionId: 'DataHub.transform',
-  metadata: {
-    inLicenseAllowed: true,
-    isTerminal: false,
-    isDataOnly: false,
-    hasArguments: true,
-    supportedEvents: [
-      BehaviorPolicyTransitionEvent.EVENT_ON_ANY,
-      BehaviorPolicyTransitionEvent.MQTT_ON_INBOUND_CONNECT,
-      BehaviorPolicyTransitionEvent.MQTT_ON_INBOUND_PUBLISH,
-      BehaviorPolicyTransitionEvent.MQTT_ON_INBOUND_SUBSCRIBE,
-      BehaviorPolicyTransitionEvent.MQTT_ON_INBOUND_DISCONNECT,
-      BehaviorPolicyTransitionEvent.CONNECTION_ON_DISCONNECT,
-    ],
-  },
-  schema: {
-    title: 'Transformation',
-    description:
-      'The list of Javascript functions used in this transformation operation. Add them directly on the graph',
-    properties: {
-      transform: {
-        type: 'array',
-        title: 'Execution order',
-        description: 'Change the order in which the transform functions will be executed',
-        items: {
-          type: 'string',
-          title: 'Function name',
-        },
-      },
-    },
-  },
-}
-
+/**
+ * @deprecated Use the fetcher-based hook instead
+ * @see useFilteredFunctionsFetcher
+ * @param type
+ * @param transition
+ */
 export const useGetFilteredFunction = (
   type: DataHubNodeType = DataHubNodeType.DATA_POLICY,
   transition?: BehaviorPolicyTransitionEvent
